Downgradable geometry (#40)
With this functionality, a user can continue to remove points until they're just left with a single point (either a `Point` or `MultiPoint` feature). I call this "downgradable" geometries.

Of course, a consuming application can always choose to reject the edit if, for instance, a position is attempting to be removed from a triangle `Polygon` (as that would downgrade it to a `LineString`).