Create EditableH3ClusterLayer && Refactor Edit Modes to support updateTentativeFeature edit type (#483)
**Refactor Edit Modes to support updateTentativeFeature edit type**:

This change refactors the following `GeoJsonEditMode`'s to support the new `updateTentativeFeature` edit type:

- Draw90DegreePolygonMode
- DrawPolygonMode (and all extended modes)
- DrawPointMode
- GeoJsonEditMode
- TwoClickPolygonMode (and all extended modes)
- ThreeClickPolygonMode (and all extended modes)

`updateTentativeFeature` is a new edit type meant to be fired whenever a polygonal shape is near completion, and the user continues to drag the cursor mid-editing. It fires on pointer move. The tentative feature created is different for each mode. Tested for all of the above modes.

This edit type will be used to update a cluster of hexes on pointer move in an `EditableH3ClusterLayer`.

**Create EditableH3ClusterLayer**:

This creates a new `EditableH3ClusterLayer`. This layer uses the `EditableGeoJsonLayer` to support editing and display of H3 hexagons via the Deck.gl `H3ClusterLayer`. This layer supports the `union`, `difference`, and `intersection` draw modes, and supports single hexagon interactions. This layer also supports selection of different hexagon clusters.
Example usage: https://imgur.com/2JGBbZS